[step4] HttpBody는 바이트스트림을 저장하는 String만 갖고, RequestManager에서 KeyValueHttpBody로 변경해서 해석하도록 변경
